Remove espanso config
I believe that my previous use of Espanso (currently not in use) was done
before the 2.x release of Espanso. Since I had not been using Epanso for
a while, I am removing the configurations and will look at re-adding
them *if* I decide to use Espanso in the future.